Move JobRegistry to a separate module
