+++ build.xml
@@ -94,6 +94,7 @@
         <ant dir="runtime/etiao" target="@{target}" useNativeBasedir="true"/>
         <ant dir="runtime/jmxcontrol" target="@{target}" useNativeBasedir="true"/>
         <ant dir="runtime/jsoncontrol" target="@{target}" useNativeBasedir="true"/>
+        <ant dir="runtime/jobregistry" target="@{target}" useNativeBasedir="true"/>
         <ant dir="utils/metrics" target="@{target}" useNativeBasedir="true"/>
 
         <ant dir="providers/direct" target="@{target}" useNativeBasedir="true"/>
