+++ runtime/.classpath
@@ -1,6 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <classpath>
 	<classpathentry kind="src" path="appservice/src/main/java"/>
+	<classpathentry kind="src" path="jobregistry/src/main/java"/>
+	<classpathentry kind="src" path="jobregistry/src/test/java"/>
 	<classpathentry kind="src" path="etiao/src/main/java"/>
 	<classpathentry kind="src" path="etiao/src/test/java"/>
 	<classpathentry kind="src" path="jmxcontrol/src/main/java"/>
