+++ runtime/jobregistry/src/main/java/quarks/runtime/jobregistry/JobEvents.java
@@ -16,10 +16,12 @@ KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License.
 */
-package quarks.execution.services;
+package quarks.runtime.jobregistry;
 
 import quarks.execution.Job;
-import quarks.execution.services.JobRegistryService.EventType;
+import quarks.execution.services.RuntimeServices;
+import quarks.execution.services.job.JobRegistryService;
+import quarks.execution.services.job.JobRegistryService.EventType;
 import quarks.function.BiConsumer;
 import quarks.function.BiFunction;
 import quarks.function.Consumer;
@@ -38,11 +40,26 @@ import quarks.topology.Topology;
 public class JobEvents {
 
     /**
-     * Generates a stream of job events.
+     * Declares a stream populated by {@link JobRegistryService} events.
+     * <p>
+     * The job registry is passed as a runtime service. At startup 
+     * {@code JobRegistryService#addListener()} is called by the 
+     * runtime to subscribe an event listener.  The listener invokes the given 
+     * {@code wrapper} function to construct a tuple from a job event
+     * and submits the tuple on the returned stream.</p>
+     * <p>
+     * When the topology's execution is terminated, 
+     * {@code JobRegistryServic#removeListener()}  in invoked to unsubscribe 
+     * the tuple source from the job registry. 
+     * </p>
      * 
-     * @param topic Topic to publish to.
-     * @param streamType Type of objects on the stream.
-     * @return sink element representing termination of this stream.
+     * @param topology the stream topology
+     * @param wrapper constructs a tuple from a job event
+     * @return new stream containing the tuples generated by the specified {@code wrapper}.
+     * 
+     * @see Topology#getRuntimeServiceSupplier() 
+     * @see JobRegistryService#addListener(BiConsumer)
+     * @see JobRegistryService#removeListener(BiConsumer)
      */
     public static <T> TStream<T> source(
             Topology topology,
