+++ runtime/jobregistry/src/main/java/quarks/runtime/jobregistry/JobRegistry.java
@@ -29,8 +29,8 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import quarks.execution.Job;
-import quarks.execution.services.JobRegistryService;
 import quarks.execution.services.ServiceContainer;
+import quarks.execution.services.job.JobRegistryService;
 import quarks.function.BiConsumer;
 
 /**
