+++ samples/topology/src/main/java/quarks/samples/topology/JobEventsSample.java
@@ -28,9 +28,9 @@ import java.util.concurrent.TimeUnit;
 import com.google.gson.JsonObject;
 
 import quarks.execution.Job;
-import quarks.execution.services.JobEvents;
-import quarks.execution.services.JobRegistryService;
+import quarks.execution.services.job.JobRegistryService;
 import quarks.providers.direct.DirectProvider;
+import quarks.runtime.jobregistry.JobEvents;
 import quarks.runtime.jobregistry.JobRegistry;
 import quarks.topology.TStream;
 import quarks.topology.Topology;
